Sectors
created a sector system in which the puzzle is split into a desired number of sectors. When a sector is solved it will go back to the main puzzle and the user will be able to choose another sector to solve. In the future the system that keeps track of the current sector might benefit from a change.