Backend for showing tiles on larger board.
This this in a very complicated and inefficient way. Next I will completely reorganize the code and make it so the tile manage aren't singletons. The sector manager will hold as many tile managers as there are sectors and each tile manager will have its own texture and values, instead of having a lot of lists in the tile manager. The sector manager will be a singleton in the end that holds all the tile manager information and can store it that way if needed.